refactor: 근태 통계 조회 queryDSL 이용하도록 수정

+++ src/main/java/com/mmp/beacon/commute/query/response/CommuteStatisticsResponse.java
@@ -12,10 +12,10 @@ public record CommuteStatisticsResponse(
     }
 
     public record CommuteStatistics(
-            int presentDays,
-            int lateDays,
-            int absentDays,
-            int totalDays
+            long presentDays,
+            long lateDays,
+            long absentDays,
+            long totalDays
     ) {
     }
 }
